Moved time parsing/formatting related components from time_conversion to common/time.
